Smattering of type fixes.

+++ org/antlr/v4/runtime/BufferedTokenStream.h
@@ -100,7 +100,7 @@ namespace org {
 
                     virtual void seek(int index) override;
 
-                    virtual int size() override;
+                    virtual size_t size() override;
                     virtual void consume() override;
 
                     /// <summary>
